// src/modules/layout/layout-options.interface.ts
export interface LayoutOptions {
    pageTitle?: string;
    pageHeading?: string;
    showSystemBar?: boolean;
    showAppBar?: boolean;
    showSideNav?: boolean;
    showMainAsFlex?: boolean;
    showContentInContainer?: boolean;
    showContentContainerIsFluid?: boolean;
    showContentContainerFillsHeight?: boolean;
    showHeader?: boolean;
    showHeaderActions?: boolean;
    showNavigation?: boolean;
    showFooter?: boolean;
    showCustomSideNav?: boolean;
}